# PaperLab/Docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PaperLab API",
    "version": "0.1.0"
  },
  "paths": {
    "/chat/state": {
      "get": {
        "tags": [
          "chat"
        ],
        "summary": "Get Chat State",
        "operationId": "get_chat_state_chat_state_get",
        "parameters": [
          {
            "name": "project_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Target project identifier",
              "title": "Project Id"
            },
            "description": "Target project identifier"
          },
          {
            "name": "thread_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Graph thread identifier",
              "title": "Thread Id"
            },
            "description": "Graph thread identifier"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatStateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/chat/stream": {
      "post": {
        "tags": [
          "chat"
        ],
        "summary": "Stream Chat Run",
        "operationId": "stream_chat_run_chat_stream_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ChatCommandInput": {
        "properties": {
          "update": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ChatStateUpdate"
              },
              {
                "type": "null"
              }
            ]
          },
          "resume": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "$ref": "#/components/schemas/ChatResumeCommand"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resume"
          }
        },
        "type": "object",
        "title": "ChatCommandInput",
        "description": "Frontend command payload used for guidance injection and resume."
      },
      "ChatInterruptPayload": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "value": {
            "additionalProperties": true,
            "type": "object",
            "title": "Value"
          }
        },
        "type": "object",
        "required": [
          "id",
          "value"
        ],
        "title": "ChatInterruptPayload",
        "description": "Serializable interrupt record returned to the frontend."
      },
      "ChatMessageInput": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type",
            "description": "LangChain-compatible message type",
            "default": "human"
          },
          "content": {
            "type": "string",
            "title": "Content",
            "description": "User-visible message content"
          }
        },
        "type": "object",
        "required": [
          "content"
        ],
        "title": "ChatMessageInput",
        "description": "Frontend message item accepted by the chat bridge."
      },
      "ChatMessagePayload": {
        "properties": {
          "id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Type"
          },
          "role": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Role"
          },
          "content": {
            "title": "Content"
          },
          "additional_kwargs": {
            "additionalProperties": true,
            "type": "object",
            "title": "Additional Kwargs"
          },
          "response_metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Response Metadata"
          }
        },
        "type": "object",
        "required": [
          "content"
        ],
        "title": "ChatMessagePayload",
        "description": "Serialized LangChain message returned to the desktop UI."
      },
      "ChatResumeCommand": {
        "properties": {
          "action": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Action"
          }
        },
        "type": "object",
        "title": "ChatResumeCommand",
        "description": "Opaque resume payload passed back into one interrupted graph."
      },
      "ChatRunRequest": {
        "properties": {
          "project_id": {
            "type": "string",
            "title": "Project Id",
            "description": "Target project identifier"
          },
          "thread_id": {
            "type": "string",
            "title": "Thread Id",
            "description": "Persistent graph thread identifier"
          },
          "input": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Input",
            "description": "Initial graph input, typically {messages: [...]}"
          },
          "command": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ChatCommandInput"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "project_id",
          "thread_id"
        ],
        "title": "ChatRunRequest",
        "description": "Request payload for one direct LangGraph-backed chat run."
      },
      "ChatStateResponse": {
        "properties": {
          "thread_id": {
            "type": "string",
            "title": "Thread Id"
          },
          "project_id": {
            "type": "string",
            "title": "Project Id"
          },
          "messages": {
            "items": {
              "$ref": "#/components/schemas/ChatMessagePayload"
            },
            "type": "array",
            "title": "Messages"
          },
          "interrupt": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ChatInterruptPayload"
              },
              {
                "type": "null"
              }
            ]
          },
          "next_nodes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Next Nodes"
          }
        },
        "type": "object",
        "required": [
          "thread_id",
          "project_id"
        ],
        "title": "ChatStateResponse",
        "description": "Thread snapshot returned by the FastAPI chat bridge."
      },
      "ChatStateUpdate": {
        "properties": {
          "messages": {
            "items": {
              "$ref": "#/components/schemas/ChatMessageInput"
            },
            "type": "array",
            "title": "Messages"
          }
        },
        "type": "object",
        "title": "ChatStateUpdate",
        "description": "State fragment applied before a resume command."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}